Fix #416, uiTableModelRowInserted on windows adds 2 rows instead of 1.
When LVM_INSERTITEM is called after we updated LVM_SETITEMCOUNT,
the actual count will be increased by one causing one extra item
to appear in the table and loading data from the model.
This extra item is invisible at first because it doesn't get painted but
you can click on it to confirm that it is actually there.

To fix this problem we need to call INSERTITEM first and
then update the ITEMCOUNT.

Fixes #416
Fixes andlabs/ui#369

cc @bcampbell
r? @andlabs